+++ hss_okta_flutter/CHANGELOG.md
@@ -29,3 +29,12 @@
 
 # 1.2.4
 * Fixed missing extension for export barrel
+
+# 1.2.5
+* Fixed Missing Interop tag for OktaAuth
+* Added getUserInfo call
+* Removed misleading UnImplemented Exception message
+* Updated setTokens to accept Tokens
+* Updated getTokens return type to Tokens
+* Changed setTokens from Future<void> to void
+* Changed getUserInfo return type to Map